+++ src/main/java/subway/Application.java
@@ -9,7 +9,7 @@ public class Application {
         final Scanner scanner = new Scanner(System.in);
         // TODO: 프로그램 구현
         InputView.initiate(scanner);
-        ApplicationContext applicationContext = ApplicationInitializer.injectDependencies();
+        SubwayApplication applicationContext = ApplicationInitializer.injectDependencies();
         applicationContext.run();
     }
 }
