+++ src/main/java/subway/domain/Station.java
@@ -1,12 +1,9 @@
 package subway.domain;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 public class Station {
     private String name;
-//    private List<Line> includedLine = new ArrayList<>();
 
     public Station(String name) {
         this.name = name;
